backend/requirements.txt
# hackmates-backend/requirements.txt

# FastAPI and ASGI server
fastapi==0.104.1
uvicorn[standard]==0.24.0
gunicorn==21.2.0

# Database
sqlalchemy==2.0.23
alembic==1.12.1
psycopg2-binary==2.9.7
asyncpg==0.29.0

# Authentication & Security
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
bcrypt==4.0.1

# HTTP & Requests
httpx==0.25.2
requests==2.31.0
aiohttp==3.9.0

# Data Processing & Validation
pydantic==2.5.0
pydantic-settings==2.0.3
email-validator==2.1.0

# AI & NLP Services
openai==1.3.7
transformers==4.35.2
torch==2.1.0
sentence-transformers==2.2.2
spacy==3.7.2
nltk==3.8.1
scikit-learn==1.3.2
numpy==1.24.3
pandas==2.0.3

# Resume Parsing & OCR
PyMuPDF==1.23.8
pypdf2==3.0.1
python-docx==0.8.11
pytesseract==0.3.10
Pillow==10.1.0
pdf2image==1.16.3

# Web Scraping
selenium==4.15.2
beautifulsoup4==4.12.2
lxml==4.9.3

# File Storage & Upload
boto3==1.34.0
python-magic==0.4.27

# Background Tasks & Caching
celery==5.3.4
redis==5.0.1
python-redis-lock==4.0.0

# Environment & Configuration
python-dotenv==1.0.0
pydantic-settings==2.0.3

# Development & Testing
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
faker==20.1.0

# Monitoring & Logging
sentry-sdk[fastapi]==1.38.0
python-json-logger==2.0.7

# CORS & Middleware
fastapi-cors==0.0.6

# Date & Time
python-dateutil==2.8.2

# Vector Database (optional)
pinecone-client==2.2.4
chromadb==0.4.18
faiss-cpu==1.7.4

# LinkedIn API (if using official API)
linkedin-api==2.2.0

# Utilities
python-slugify==8.0.1
phonenumbers==8.13.25
pytz==2023.3

# Production
prometheus-client==0.19.0
uvloop==0.19.0  # For better performance on Linux